refactor: VideoCard 내부 썸네일 및 메타 정보 컴포넌트로 분리 #9
