+++ youtube-ari/src/components/VideoGrid/VideoThumbnail.jsx
@@ -1,0 +1,15 @@
+const VideoThumbnail = ({ thumbnail }) => {
+  return (
+    <div className="relative w-full">
+      <div className="block aspect-video overflow-hidden rounded-xl">
+        <img
+          src={thumbnail}
+          alt="Video Thumbnail"
+          className="h-full w-full object-cover"
+        />
+      </div>
+    </div>
+  );
+};
+
+export default VideoThumbnail;
